Remove util package & add changes to appropriate classes
