+++ browser/custom_app_icons/android/java/res/layout/brave_confirmation_dialog_layout.xml
