+++ browser/custom_app_icons/android/java/res/values-night/custom_app_icons_colors.xml
