+++ browser/custom_app_icons/android/java/res/values/custom_app_icons_styles.xml
