--- browser/utils/android/BUILD.gn
@@ -1,35 +1,0 @@
-# Copyright (c) 2025 The Brave Authors. All rights reserved.
-# This Source Code Form is subject to the terms of the Mozilla Public
-# License, v. 2.0. If a copy of the MPL was not distributed with this file,
-# You can obtain one at https://mozilla.org/MPL/2.0/.
-
-import("//build/config/android/rules.gni")
-import("//build/config/locales.gni")
-import("//chrome/common/features.gni")
-import("//tools/grit/grit_rule.gni")
-
-android_library("java") {
-  sources = [
-    "java/src/org/chromium/brave/browser/utils/DrawableUtils.java",
-    "java/src/org/chromium/brave/browser/utils/confirm_dialog/BraveConfirmationDialog.java",
-    "java/src/org/chromium/brave/browser/utils/confirm_dialog/OnConfirmationDialogListener.java",
-  ]
-
-  deps = [
-    ":java_resources",
-    "//third_party/androidx:androidx_core_core_java",
-  ]
-
-  resources_package = "org.chromium.brave.browser.utils"
-}
-
-android_resources("java_resources") {
-  sources = [
-    "java/res/drawable/brave_action_button_bg.xml",
-    "java/res/drawable/confirmation_dialog_shape.xml",
-    "java/res/layout/brave_confirmation_dialog_layout.xml",
-    "java/res/values-night/brave_android_utils_colors.xml",
-    "java/res/values/brave_android_utils_colors.xml",
-    "java/res/values/brave_android_utils_styles.xml",
-  ]
-}
